Move redshift queries
