+++ ingestion/src/metadata/ingestion/source/database/redshift/lineage.py
@@ -31,10 +31,10 @@ workflowConfig:
 """
 
 from metadata.ingestion.source.database.lineage_source import LineageSource
+from metadata.ingestion.source.database.redshift.queries import REDSHIFT_SQL_STATEMENT
 from metadata.ingestion.source.database.redshift.query_parser import (
     RedshiftQueryParserSource,
 )
-from metadata.utils.sql_queries import REDSHIFT_SQL_STATEMENT
 
 
 class RedshiftLineageSource(RedshiftQueryParserSource, LineageSource):
