+++ ingestion/src/metadata/utils/sql_queries.py
@@ -14,185 +14,6 @@ SQL Queries used during ingestion
 
 import textwrap
 
-# Not able to use SYS_QUERY_HISTORY here. Few users not getting any results
-REDSHIFT_SQL_STATEMENT = textwrap.dedent(
-    """
-  WITH
-  queries AS (
-    SELECT *
-      FROM pg_catalog.stl_query
-     WHERE userid > 1
-          {filters}
-          -- Filter out all automated & cursor queries
-          AND querytxt NOT LIKE '/* {{"app": "OpenMetadata", %%}} */%%'
-          AND querytxt NOT LIKE '/* {{"app": "dbt", %%}} */%%'
-          AND aborted = 0
-          AND starttime >= '{start_time}'
-          AND starttime < '{end_time}'
-          LIMIT {result_limit}
-  ),
-  full_queries AS (
-    SELECT
-          query,
-          LISTAGG(CASE WHEN LEN(RTRIM(text)) = 0 THEN text ELSE RTRIM(text) END, '')
-            WITHIN GROUP (ORDER BY sequence) AS query_text
-      FROM pg_catalog.stl_querytext
-      WHERE sequence < 327	-- each chunk contains up to 200, RS has a maximum str length of 65535.
-    GROUP BY query
-  ),
-  raw_scans AS (
-    -- We have one row per table per slice so we need to get rid of the dupes
-    SELECT distinct query, tbl
-      FROM pg_catalog.stl_scan
-  ),
-  scans AS (
-  	SELECT DISTINCT
-  		query,
-  		sti.database AS database_name,
-      sti.schema AS schema_name
-  	  FROM raw_scans AS s
-          INNER JOIN pg_catalog.svv_table_info AS sti
-            ON (s.tbl)::oid = sti.table_id
-  )
-  SELECT DISTINCT
-        q.userid,
-        s.query AS query_id,
-        RTRIM(u.usename) AS user_name,
-        fq.query_text,
-        s.database_name,
-        s.schema_name,
-        q.starttime AS start_time,
-        q.endtime AS end_time,
-        q.aborted AS aborted
-    FROM scans AS s
-        INNER JOIN queries AS q
-          ON s.query = q.query
-        INNER JOIN full_queries AS fq
-          ON s.query = fq.query
-        INNER JOIN pg_catalog.pg_user AS u
-          ON q.userid = u.usesysid
-    ORDER BY q.endtime DESC
-"""
-)
-
-
-REDSHIFT_GET_ALL_RELATION_INFO = textwrap.dedent(
-    """
-        SELECT
-          c.relkind,
-          n.oid as "schema_oid",
-          n.nspname as "schema",
-          c.oid as "rel_oid",
-          c.relname,
-          CASE c.reldiststyle
-            WHEN 0 THEN 'EVEN' WHEN 1 THEN 'KEY' WHEN 8 THEN 'ALL' END
-            AS "diststyle",
-          c.relowner AS "owner_id",
-          u.usename AS "owner_name",
-          TRIM(TRAILING ';' FROM pg_catalog.pg_get_viewdef(c.oid, true))
-            AS "view_definition",
-          pg_catalog.array_to_string(c.relacl, '\n') AS "privileges"
-        FROM pg_catalog.pg_class c
-             LEFT JOIN pg_catalog.pg_namespace n ON n.oid = c.relnamespace
-             JOIN pg_catalog.pg_user u ON u.usesysid = c.relowner
-        WHERE c.relkind IN ('r', 'v', 'm', 'S', 'f')
-          AND n.nspname !~ '^pg_'
-        ORDER BY c.relkind, n.oid, n.nspname;
-        """
-)
-
-
-REDSHIFT_GET_SCHEMA_COLUMN_INFO = textwrap.dedent(
-    """
-            SELECT
-              n.nspname as "schema",
-              c.relname as "table_name",
-              att.attname as "name",
-              format_encoding(att.attencodingtype::integer) as "encode",
-              format_type(att.atttypid, att.atttypmod) as "type",
-              att.attisdistkey as "distkey",
-              att.attsortkeyord as "sortkey",
-              att.attnotnull as "notnull",
-              pg_catalog.col_description(att.attrelid, att.attnum)
-                as "comment",
-              adsrc,
-              attnum,
-              pg_catalog.format_type(att.atttypid, att.atttypmod),
-              pg_catalog.pg_get_expr(ad.adbin, ad.adrelid) AS DEFAULT,
-              n.oid as "schema_oid",
-              c.oid as "table_oid"
-            FROM pg_catalog.pg_class c
-            LEFT JOIN pg_catalog.pg_namespace n
-              ON n.oid = c.relnamespace
-            JOIN pg_catalog.pg_attribute att
-              ON att.attrelid = c.oid
-            LEFT JOIN pg_catalog.pg_attrdef ad
-              ON (att.attrelid, att.attnum) = (ad.adrelid, ad.adnum)
-            WHERE n.nspname !~ '^pg_'
-              AND att.attnum > 0
-              AND NOT att.attisdropped
-              {schema_clause}
-            UNION
-            SELECT
-              view_schema as "schema",
-              view_name as "table_name",
-              col_name as "name",
-              null as "encode",
-              col_type as "type",
-              null as "distkey",
-              0 as "sortkey",
-              null as "notnull",
-              null as "comment",
-              null as "adsrc",
-              null as "attnum",
-              col_type as "format_type",
-              null as "default",
-              null as "schema_oid",
-              null as "table_oid"
-            FROM pg_get_late_binding_view_cols() cols(
-              view_schema name,
-              view_name name,
-              col_name name,
-              col_type varchar,
-              col_num int)
-            WHERE 1 {schema_clause}
-            UNION
-            SELECT schemaname AS "schema",
-               tablename AS "table_name",
-               columnname AS "name",
-               null AS "encode",
-               -- Spectrum represents data types differently.
-               -- Standardize, so we can infer types.
-               CASE
-                 WHEN external_type = 'int' THEN 'integer'
-                 ELSE
-                   replace(
-                    replace(external_type, 'decimal', 'numeric'),
-                    'varchar', 'character varying')
-                 END
-                    AS "type",
-               null AS "distkey",
-               0 AS "sortkey",
-               null AS "notnull",
-               null AS "comment",
-               null AS "adsrc",
-               null AS "attnum",
-               CASE
-                 WHEN external_type = 'int' THEN 'integer'
-                 ELSE
-                   replace(
-                    replace(external_type, 'decimal', 'numeric'),
-                    'varchar', 'character varying')
-                 END
-                    AS "format_type",
-               null AS "default",
-               null AS "schema_oid",
-               null AS "table_oid"
-            FROM svv_external_columns
-            ORDER BY "schema", "table_name", "attnum";
-            """
-)
-
 SNOWFLAKE_SQL_STATEMENT = textwrap.dedent(
     """
         SELECT
@@ -455,13 +276,6 @@ SNOWFLAKE_GET_CLUSTER_KEY = """
   from   information_schema.tables 
   where  TABLE_TYPE = 'BASE TABLE'
   and CLUSTERING_KEY is not null
-"""
-
-
-REDSHIFT_PARTITION_DETAILS = """
-  select "schema", "table", diststyle
-  from SVV_TABLE_INFO
-  where diststyle not like 'AUTO%%'
 """
